fix: delete imageService
