--- packages/image/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "@wrtnlabs/connector-image",
-  "version": "0.0.1",
-  "description": "",
-  "main": "src/index.ts",
-  "exports": {
-    "import": "./lib/index.mjs",
-    "require": "./lib/index.cjs",
-    "default": "./lib/index.cjs"
-  },
-  "scripts": {
-    "prepare": "ts-patch install && typia patch",
-    "build": "pnpm build:main && pnpm build:test",
-    "build:main": "rimraf lib && tsc && rollup -c",
-    "build:test": "rimraf bin && tsc -p test/tsconfig.json",
-    "dev": "pnpm build:test --watch",
-    "eslint": "eslint ./**/*.ts",
-    "test": "node bin/test"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "MIT",
-  "devDependencies": {
-    "@eslint/js": "^9.20.0",
-    "@nestia/e2e": "^0.7.0",
-    "@rollup/plugin-terser": "^0.4.4",
-    "@rollup/plugin-typescript": "^12.1.1",
-    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
-    "chalk": "^4.1.2",
-    "dotenv": "^16.3.1",
-    "dotenv-expand": "^10.0.0",
-    "eslint": "^9.20.1",
-    "path": "^0.12.7",
-    "prettier": "^3.3.2",
-    "rimraf": "^6.0.1",
-    "rollup": "^4.34.8",
-    "ts-node": "^10.9.2",
-    "ts-patch": "^3.3.0",
-    "tstl": "^3.0.0",
-    "typescript": "~5.7.3",
-    "typescript-eslint": "^8.24.1",
-    "typescript-transform-paths": "^3.5.1"
-  },
-  "dependencies": {
-    "axios": "^1.8.1",
-    "sharp": "^0.33.5",
-    "typia": "^8.0.1"
-  }
-}
