--- packages/image/src/index.ts
@@ -1,5 +1,0 @@
-// TYPES
-export * from "./structures/IImageService";
-
-// CLASS
-export * from "./image/ImageService";
