--- packages/image/test/index.ts
@@ -1,45 +1,0 @@
-import { DynamicExecutor } from "@nestia/e2e";
-import chalk from "chalk";
-
-import { TestGlobal } from "./TestGlobal";
-
-const main = async (): Promise<void> => {
-  // DO TEST
-  const include: string[] = TestGlobal.getArguments("include");
-  const exclude: string[] = TestGlobal.getArguments("exclude");
-  const report: DynamicExecutor.IReport = await DynamicExecutor.validate({
-    prefix: "test_",
-    location: __dirname + "/features",
-    parameters: () => [],
-    onComplete: (exec) => {
-      const trace = (str: string) =>
-        console.log(`  - ${chalk.green(exec.name)}: ${str}`);
-      if (exec.error === null) {
-        const elapsed: number =
-          new Date(exec.completed_at).getTime() -
-          new Date(exec.started_at).getTime();
-        trace(`${chalk.yellow(elapsed.toLocaleString())} ms`);
-      } else trace(chalk.red(exec.error.name));
-    },
-    filter: (name) =>
-      (include.length ? include.some((str) => name.includes(str)) : true) &&
-      (exclude.length ? exclude.every((str) => !name.includes(str)) : true),
-  });
-
-  const exceptions: Error[] = report.executions
-    .filter((exec) => exec.error !== null)
-    .map((exec) => exec.error!);
-  if (exceptions.length === 0) {
-    console.log("Success");
-    console.log("Elapsed time", report.time.toLocaleString(), `ms`);
-  } else {
-    for (const exp of exceptions) console.log(exp);
-    console.log("Failed");
-    console.log("Elapsed time", report.time.toLocaleString(), `ms`);
-    process.exit(-1);
-  }
-};
-main().catch((error) => {
-  console.error(error);
-  process.exit(-1);
-});
