--- packages/image/test/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "target": "ES2015",
-    "outDir": "../bin",
-    "paths": {
-      "@wrtnlabs/connector-image": ["../src/index.ts"],
-      "@wrtnlabs/connector-image/lib/*": ["../src/*"]
-    },
-    "plugins": [
-      { "transform": "typescript-transform-paths" },
-      { "transform": "typia/lib/transform" }
-    ]
-  },
-  "include": ["../src", "./"]
-}
