+++ packages/shared/src/structures/connector/index.ts
@@ -1,1 +1,2 @@
 export * from "./spreadsheet";
+export * from "./image";
