+++ packages/shared/src/structures/index.ts
@@ -1,4 +1,3 @@
 export * from "./common";
 export * from "./connector";
-export * from "./internal";
 export * from "./types";
