+++ packages/shared/src/utils/index.ts
@@ -11,6 +11,7 @@ export * from "./PaginationUtil";
 export * from "./retry";
 export * from "./TypedSplit";
 export * from "./bufferUtil";
+export * from "./imageUtil";
 
 // CONSTANTS
 export * from "./constants/extensions";
